requirements.txt
opencv-python
scipy
imutils
numpy
dlib
face_recognition
twilio
python-pushover
RPi.GPIO